+++ InstallerLib/InstalledCheck.cs
@@ -81,6 +81,8 @@ namespace InstallerLib
                 l_check = new InstalledCheckRegistry();
             else if (element.Attributes["type"].InnerText == "check_product")
                 l_check = new InstalledCheckProduct();
+            else if (element.Attributes["type"].InnerText == "check_wmi")
+                l_check = new InstalledCheckWmi();
             else
                 throw new Exception(string.Format(
                     "Invalid installcheck type: {0}", element.Attributes["type"].InnerText));