+++ dotNetInstallerToolsLib/WmiUtil.h
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace DVLib
+{
+	class WmiUtil
+	{
+	public:
+		static bool Any(const std::wstring& query);
+	};
+}