+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Features
 * [#114](https://github.com/dotnetinstaller/dotnetinstaller/issues/114) - Added support for operating system product type filters - [@icnocop](https://github.com/icnocop).
 * [#153](https://github.com/dotnetinstaller/dotnetinstaller/issues/153) - Added support for sub-directory installed check comparison using semantic versioning - [@icnocop](https://github.com/icnocop).
 * [#156](https://github.com/dotnetinstaller/dotnetinstaller/pull/156) - Building and distributing 64-bit version of CabLib.dll for reference by 64-bit processes - [@icnocop](https://github.com/icnocop).
+* [#157](https://github.com/dotnetinstaller/dotnetinstaller/issues/157) - Added Installed Check Type which queries WMI - [@icnocop](https://github.com/icnocop).
 
 3.2.115 (1/15/2022)
 ===================